Added error checking for non-csv file types, as well as ensuring that parsed csv data fits to ModelObject type
