+++ package.json
@@ -23,6 +23,7 @@
     "antd": "^3.26.7",
     "axios": "^0.19.1",
     "classnames": "^2.2.6",
+    "csvtojson": "^2.0.10",
     "dom4": "^2.1.5",
     "jquery": "^3.4.1",
     "node-sass": "^4.13.1",
